% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.env.R
\name{add.env}
\alias{add.env}
\title{Takes a set of points, a buffer radius, a sample size, and a mask and returns
randomly sampled points from within that buffer radius.
Code modified from Elith and Hijmans SDM with R tutorial}
\usage{
add.env(species, env, verbose = TRUE)
}
\arguments{
\item{species}{An enmtools.species or enmtools.clade object}

\item{env}{A raster stack of environmental variables}
}
\description{
Takes a set of points, a buffer radius, a sample size, and a mask and returns
randomly sampled points from within that buffer radius.
Code modified from Elith and Hijmans SDM with R tutorial
}
