% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.overlap.correlation.R
\name{age.overlap.correlation}
\alias{age.overlap.correlation}
\title{Takes an overlap matrix and a tree and runs permutation tests to
determine the statistical significance of the relationship between
overlap and time}
\usage{
age.overlap.correlation(overlap, tree, nreps)
}
\arguments{
\item{overlap}{An overlap matrix}

\item{tree}{A tree}

\item{nreps}{A number of reps to do}
}
\description{
Takes an overlap matrix and a tree and runs permutation tests to
determine the statistical significance of the relationship between
overlap and time
}

