% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.depends.R
\name{assert.extras}
\alias{assert.extras}
\title{Assert that the extra packages needed for an ENMTools function are installed and available}
\usage{
assert.extras(funs = NULL)
}
\arguments{
\item{funs}{character vector of ENMTools function names to install extra dependencies for. If `NULL`, extras for all ENMTools functions will be installed.}
}
\value{
None, this function is used for its side-effects
}
\description{
Triggers an error if any of the extra packages required for an ENMTools function are not available.
}
\examples{
if(check.extras("enmtools.gam")) {
  assert.extras("enmtools.gam")
}
}
