% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.points.buffer.R
\name{background.points.buffer}
\alias{background.points.buffer}
\title{Takes a set of points, a buffer radius, a sample size, and a mask and returns
randomly sampled points from within that buffer radius.
Code modified from Elith and Hijmans SDM with R tutorial}
\usage{
background.points.buffer(points, radius, n, mask)
}
\arguments{
\item{points}{A two column data frame with X and Y coordinates}

\item{radius}{Radius for circular buffers to draw around points, in meters.}

\item{n}{Sample size for number of background points to return}

\item{mask}{A raster to use as a mask for drawing points}
}
\value{
A data frame of points drawn at random from circular buffers around the occurrence points.
}
\description{
Takes a set of points, a buffer radius, a sample size, and a mask and returns
randomly sampled points from within that buffer radius.
Code modified from Elith and Hijmans SDM with R tutorial
}
