% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.shape.buffer.R
\name{background.shape.buffer}
\alias{background.shape.buffer}
\title{Takes a set of points and a buffer radius, and returns
a polygon.
Code modified from Elith and Hijmans SDM with R tutorial}
\usage{
background.shape.buffer(points, radius)
}
\arguments{
\item{points}{A two column data frame with X and Y coordinates}

\item{radius}{Radius for circular buffers to draw around points, in meters.}
}
\value{
A polygon shapefile.
}
\description{
Takes a set of points and a buffer radius, and returns
a polygon.
Code modified from Elith and Hijmans SDM with R tutorial
}
