% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.B1.R
\name{calc.B1}
\alias{calc.B1}
\title{calc.B1, Calculates standardized version of Levins (1968) B1 measure of niche breadth given a vector of suitabilities}
\usage{
calc.B1(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
B1 A calculation of Levins (1968) B1 metric
}
\description{
calc.B1, Calculates standardized version of Levins (1968) B1 measure of niche breadth given a vector of suitabilities
}
\examples{
calc.B1(c(1, .001, .001))
}
\keyword{breadth}
\keyword{enm}
\keyword{niche}
\keyword{sdm}
