% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.B2.R
\name{calc.B2}
\alias{calc.B2}
\title{calc.B2, Calculates standardized version of Levins (1968) B2 measure of niche breadth given a vector of suitabilities}
\usage{
calc.B2(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
B2 A calculation of Levins (1968) B2 metric
}
\description{
calc.B2, Calculates standardized version of Levins (1968) B2 measure of niche breadth given a vector of suitabilities
}
\examples{
calc.B2(c(1, .001, .001))
}
\keyword{breadth}
\keyword{enm}
\keyword{niche}
\keyword{sdm}
