% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.clade.R
\name{check.clade}
\alias{check.clade}
\title{Checking for complians of an enmtools.clade object}
\usage{
check.clade(this.clade)
}
\arguments{
\item{this.clade}{An enmtools.clade object}
}
\description{
Checking for complians of an enmtools.clade object
}

