% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.env.R
\name{check.env}
\alias{check.env}
\title{Automating some basic tasks for using a raster stack for modeling.  Checks rasters for same extent and resolution, and sets values in each layer to NA if there is an NA in any other layer.}
\usage{
check.env(env, verbose = FALSE)
}
\arguments{
\item{env}{A stack of environmental rasters}

\item{verbose}{Controls printing of progress messages}
}
\value{
A raster stack.
}
\description{
Automating some basic tasks for using a raster stack for modeling.  Checks rasters for same extent and resolution, and sets values in each layer to NA if there is an NA in any other layer.
}
\examples{
data(euro.worldclim)
check.env(euro.worldclim)
}
