% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.species.R
\name{combine.species}
\alias{combine.species}
\title{Takes a list of enmtools.species objects and combines them into a single enmtools.species object}
\usage{
combine.species(species.list)
}
\arguments{
\item{species.list}{List of enmtools.species objects that you want to combine together}
}
\description{
Takes a list of enmtools.species objects and combines them into a single enmtools.species object
}
