% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{domain_bridge}
\alias{domain_bridge}
\title{Wrapper function for \code{dismo::domain()}}
\usage{
domain_bridge(x, y)
}
\arguments{
\item{x}{Matrix or data.frame of environmental variables at points}

\item{y}{Single column matrix or data.frame containing a two-level factor, specifying whether each point is a 'presence' point or a 'background' point.}
}
\value{
The result of calling \code{\link[dismo:domain]{dismo::domain()}}
}
\description{
Wraps \code{\link[dismo:domain]{dismo::domain()}} for use in a \code{parsnip} model specification.
Mostly for internal use. Exported so it works properly with parallel computation in \code{tidymodels}
}
\examples{
dat <- enmtools.prep(iberolacerta.clade$species$monticola, euro.worldclim)$data
domain_bridge(dat[ , c("bio1", "bio9")], dat[, "presence", drop = FALSE])
}
