% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.species.R
\name{drop.species}
\alias{drop.species}
\title{Takes a an ENMTools clade object and a vector of species names.  Drops the species from the tree and removes data from the clade object.}
\usage{
drop.species(clade, species)
}
\arguments{
\item{clade}{An ENMTools clade object}

\item{species}{A name or vector of names of species within the enmtools.clade object.}
}
\description{
Takes a an ENMTools clade object and a vector of species names.  Drops the species from the tree and removes data from the clade object.
}
\examples{
data(iberolacerta.clade)
new.clade <- drop.species(iberolacerta.clade, c("cyreni", "monticola"))
}
