% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.precheck.R
\name{enm.precheck}
\alias{enm.precheck}
\title{Takes an emtools.species object and does tests preparatory to building an ENM}
\usage{
enm.precheck(species)
}
\arguments{
\item{species}{An enmtools.species object}
}
\description{
Takes an emtools.species object and does tests preparatory to building an ENM
}

