% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.anw.sdm.R
\name{enmtools.anw.sdm}
\alias{enmtools.anw.sdm}
\title{Takes an emtools.species object with presence and background points, and builds an Available Niche Weighted SDM (ANWSDM)}
\usage{
enmtools.anw.sdm(species, env, f = NULL, method = "glm",
  standardise = TRUE, test.prop = 0, eval = TRUE, nback = 1000, k = 4,
  report = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{method}{SDM method to use. Currently, only 'gam' and 'glm' are implemented.}

\item{standardise}{Should environmental data be standardised before model fitting? (Recommended)}

\item{test.prop}{Proportion of data to withhold for model evaluation}

\item{eval}{Determines whether model evaluation should be done.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{k}{Dimension of the basis used to represent the smooth terms if \code{method="gam"}. Ignored otherwise. See documentation for \code{\link[mgcv]{s}} for details.}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{...}{Arguments to be passed to \code{method} function. e.g. parameter \code{k} to be passed to}

\item{formula}{Standard formula applicable to \code{method}}
}
\description{
Takes an emtools.species object with presence and background points, and builds an Available Niche Weighted SDM (ANWSDM)
}
