% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.bc.R
\name{enmtools.bc}
\alias{enmtools.bc}
\title{Takes an emtools.species object with presence and background points, and builds a Bioclim model}
\usage{
enmtools.bc(species, env = NA, test.prop = 0, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{test.prop}{Proportion of data to withhold for model evaluation}

\item{...}{Arguments to be passed to bioclim()}
}
\description{
Takes an emtools.species object with presence and background points, and builds a Bioclim model
}

