% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.calibrate.R
\name{enmtools.calibrate}
\alias{enmtools.calibrate}
\title{Takes an emtools.model object, and reformats it to run through the CalibratR package, calculates Continuous Boyce Index, and runs a Hosmer-Lemeshow goodness-of-fit test.  Can either do a full CalibratR run or just return ECE/MCE statistics and plots.}
\usage{
enmtools.calibrate(model, recalibrate = FALSE, cuts = 11, ...)
}
\arguments{
\item{model}{An enmtools.model object}

\item{recalibrate}{When TRUE, does a full CalibratR "calibrate" run to recalibrate the model.  When FALSE, just returns metrics and plots measuring calibration of the model as is.}

\item{cuts}{The number of bins to split suitability scores into for calculating calibration.}

\item{...}{Further arguments to be passed to CalibratR's "calibrate" function.}
}
\description{
Takes an emtools.model object, and reformats it to run through the CalibratR package, calculates Continuous Boyce Index, and runs a Hosmer-Lemeshow goodness-of-fit test.  Can either do a full CalibratR run or just return ECE/MCE statistics and plots.
}
\examples{
data(euro.worldclim)
data(iberolacerta.clade)
monticola.glm <- enmtools.glm(iberolacerta.clade$species$monticola,
env = euro.worldclim, f = pres ~ bio1 + bio9, test.prop = 0.3)
enmtools.calibrate(monticola.glm)
}
