% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.ecospat.id.R
\name{enmtools.ecospat.id}
\alias{enmtools.ecospat.id}
\title{enmtools.ecospat.id, Runs an ecospat identity test using enmtool.species objects.}
\usage{
enmtools.ecospat.id(species.1, species.2, env, nreps = 99, layers = NULL,
  th.sp = 0, th.env = 0, R = 100, ...)
}
\arguments{
\item{species.1}{An enmtools.species object}

\item{species.2}{An enmtools.species object}

\item{env}{A set of environmental layers}

\item{nreps}{The number of pseudoreplicates to perform}

\item{layers}{A vector of length 2 containing the names of the layers to be used.}

\item{th.sp}{Quantile of species densities used as a threshold to exclude low species density values.  See documentation for ecospat.grid.clim.dyn.}

\item{th.env}{Quantile of environmental densities across studye sites used as threshold to exclude low
environmental density values.  See documentation for ecospat.grid.clim.dyn.}

\item{R}{Resolution of the grid. See documentation for ecospat.grid.clim.dyn.}
}
\value{
results Some results, once I figure out what results to return
}
\description{
enmtools.ecospat.id, Runs an ecospat identity test using enmtool.species objects.
}
\examples{
enmtools.ecospat.id(ahli, allogus)
}
\keyword{enm}
\keyword{niche}
\keyword{plot}
\keyword{sdm}
