% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.evaluate.R
\name{enmtools.evaluate}
\alias{enmtools.evaluate}
\title{Takes an enmtools model object and an emtools.species object with presence and background points, formats it, and calls dismo evaluate function}
\usage{
enmtools.evaluate(species, model, env, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A RasterStack object containing the environmental predictors}

\item{...}{Arguments to be passed to evaluate()}

\item{species}{An enmtools model object (enmtools.glm, enmtools.dm, etc.)}
}
\description{
Takes an enmtools model object and an emtools.species object with presence and background points, formats it, and calls dismo evaluate function
}
