% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.hypervolume.R
\name{enmtools.hypervolume}
\alias{enmtools.hypervolume}
\title{Takes an enmtools.species object and builds a hypervolume model}
\usage{
enmtools.hypervolume(
  species,
  env,
  f = NULL,
  test.prop = 0,
  eval = TRUE,
  nback = 1000,
  env.nback = 10000,
  report = NULL,
  overwrite = FALSE,
  rts.reps = 0,
  bg.source = "default",
  verbose = FALSE,
  clamp = TRUE,
  corner = NA,
  bias = NA,
  method = "gaussian",
  samples.per.point = 1000,
  reduction.factor = 0.5,
  ...
)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A SpatRaster of environmental data. Note: hypervolume works best with a small number of
uncorrelated environmental variables (typically 2-6 dimensions).}

\item{f}{Standard R formula or tidymodels recipe (optional). If NULL, all environmental variables will be used.}

\item{test.prop}{Proportion of data to withhold randomly for model evaluation, or "block" for spatially structured evaluation.}

\item{eval}{Determines whether model evaluation should be done. Turned on by default.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{env.nback}{Number of points to draw from environment space for environment space discrimination metrics.}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{rts.reps}{The number of replicates to do for a Raes and ter Steege-style test of significance}

\item{bg.source}{Source for drawing background points. If "points", it just uses the background points that are already in the species object. If "range", it uses the range raster. If "env", it draws points at random from the entire study area outlined by the first environmental layer.}

\item{verbose}{Controls printing of various messages progress reports. Defaults to FALSE.}

\item{clamp}{When set to TRUE, clamps the environmental layers so that predictions made outside the min/max of the training data for each predictor are set to the value for the min/max for that predictor.}

\item{corner}{An integer from 1 to 4. Selects which corner to use for "block" test data. By default the corner is selected randomly.}

\item{bias}{An optional raster estimating relative sampling effort per grid cell. Will be used for drawing background data.}

\item{method}{Method for constructing hypervolumes: "gaussian" (default) or "svm".}

\item{samples.per.point}{Number of random samples per point for hypervolume estimation. Default 1000.}

\item{reduction.factor}{Value between 0 and 1 for prediction speed. Lower values are faster but less accurate. Default 0.5.}

\item{...}{Additional arguments to be passed to hypervolume construction functions}
}
\value{
An enmtools model object containing species name, model object, suitability raster, and any evaluation objects that were created.
}
\description{
Takes an enmtools.species object and builds a hypervolume model
}
\details{
Hypervolume models estimate the environmental niche as an n-dimensional hypervolume
using kernel density estimation. This is a presence-only method that does not use background points
for model fitting (though background points are still used for model evaluation).

Note that hypervolume models work best with a small number of uncorrelated environmental
variables (typically 2-6). Using too many variables can lead to poor performance due to the
curse of dimensionality.

The function internally standardizes environmental data before fitting the hypervolume,
and stores the scaling parameters for use in prediction.
}
\examples{
\dontrun{
# Use a small number of environmental variables
env_subset <- euro.worldclim[[c("bio1", "bio12")]]
monticola.hv <- enmtools.hypervolume(iberolacerta.clade$species$monticola, env = env_subset)
}

}
\seealso{
\code{\link[hypervolume]{hypervolume_gaussian}} for the underlying modeling function.
}
