% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.maxent.R
\name{enmtools.maxent}
\alias{enmtools.maxent}
\title{Takes an emtools.species object with presence and background points, and builds a maxent model}
\usage{
enmtools.maxent(species, env, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{...}{Arguments to be passed to maxent()}

\item{formula}{Standard R formula}
}
\description{
Takes an emtools.species object with presence and background points, and builds a maxent model
}

