% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.maxnet.R
\name{enmtools.maxnet}
\alias{enmtools.maxnet}
\title{Takes an enmtools.species object and builds a maxnet model}
\usage{
enmtools.maxnet(
  species,
  env,
  f = NULL,
  test.prop = 0,
  eval = TRUE,
  nback = 1000,
  env.nback = 10000,
  report = NULL,
  overwrite = FALSE,
  rts.reps = 0,
  weights = "equal",
  bg.source = "default",
  verbose = FALSE,
  clamp = TRUE,
  corner = NA,
  bias = NA,
  ...
)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A SpatRaster of environmental data.}

\item{f}{Standard R formula or tidymodels recipe (optional). If NULL, all environmental variables will be used.}

\item{test.prop}{Proportion of data to withhold randomly for model evaluation, or "block" for spatially structured evaluation.}

\item{eval}{Determines whether model evaluation should be done. Turned on by default.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{env.nback}{Number of points to draw from environment space for environment space discrimination metrics.}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{rts.reps}{The number of replicates to do for a Raes and ter Steege-style test of significance}

\item{weights}{If this is set to "equal", presences and background data will be assigned weights so that the sum of all presence points weights equals the sum of all background point weights. Otherwise, weights are not provided to the model.}

\item{bg.source}{Source for drawing background points. If "points", it just uses the background points that are already in the species object. If "range", it uses the range raster. If "env", it draws points at random from the entire study area outlined by the first environmental layer.}

\item{verbose}{Controls printing of various messages progress reports. Defaults to FALSE.}

\item{clamp}{When set to TRUE, clamps the environmental layers so that predictions made outside the min/max of the training data for each predictor are set to the value for the min/max for that predictor.}

\item{corner}{An integer from 1 to 4. Selects which corner to use for "block" test data. By default the corner is selected randomly.}

\item{bias}{An optional raster estimating relative sampling effort per grid cell. Will be used for drawing background data.}

\item{...}{Additional arguments to be passed to maxnet::maxnet()}
}
\value{
An enmtools model object containing species name, model formula (if any), model object, suitability raster, marginal response plots, and any evaluation objects that were created.
}
\description{
Takes an enmtools.species object and builds a maxnet model
}
\details{
maxnet provides a pure R implementation of the Maxent algorithm using glmnet for regularization.
Unlike the original Java-based Maxent, maxnet does not require Java installation.
The model uses the same feature types (linear, quadratic, product, threshold, hinge) and
regularization approach as Maxent.
}
\examples{
\dontrun{
enmtools.maxnet(iberolacerta.clade$species$monticola, env = euro.worldclim)
}

}
\seealso{
\code{\link[maxnet]{maxnet}} for the underlying modeling function.
}
