% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.ppmlasso.R
\name{enmtools.ppmlasso}
\alias{enmtools.ppmlasso}
\title{Takes an emtools.species object with presence and background points, and builds a Point Process Model (with Lasso penalty)}
\usage{
enmtools.ppmlasso(species, env, f = NULL, test.prop = 0, eval = TRUE,
  nback = 10000, report = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{test.prop}{Proportion of data to withhold for model evaluation}

\item{eval}{Determines whether model evaluation should be done.  Turned on by default, but moses turns it off to speed things up.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{...}{Arguments to be passed to ppmlasso()}

\item{formula}{Standard R formula}
}
\description{
Takes an emtools.species object with presence and background points, and builds a Point Process Model (with Lasso penalty)
}
