% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{enmtools.prep}
\alias{enmtools.prep}
\title{Prepare data for ENMTools models}
\usage{
enmtools.prep(
  x,
  env = NA,
  nback = 1000,
  bg.source = "default",
  verbose = FALSE,
  bias = NA,
  weights = "none"
)
}
\arguments{
\item{x}{An \code{enmtools.species} object}

\item{env}{A SpatRaster of environmental data.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{bg.source}{Source for drawing background points.  If "points", it just uses the background points that are already in the species object.  If "range", it uses the range raster.  If "env", it draws points at randome from the entire study area outlined by the first environmental layer.}

\item{verbose}{Controls printing of various messages progress reports.  Defaults to FALSE.}

\item{bias}{An optional raster estimating relative sampling effort per grid cell.  Will be used for drawing background data.}

\item{weights}{If this is set to "equal", presences and background data will be assigned weights so that the sum of all presence points weights equals the sum of all background point weights.  Otherwise, weights are not provided to the model.}
}
\value{
A list containing two elements: \code{data} and \code{species}. \code{data} contains a data.frame with formatted data suitable for modelling. \code{species} contains an updated \code{enmtools.species} object with background point information filled in.
}
\description{
This is mostly used internally by ENMTools but we export it for users who want to
see how the sausage is made.
}
\examples{
enmtools.prep(iberolacerta.clade$species$monticola, euro.worldclim)
}
