% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{enmtools.tidy}
\alias{enmtools.tidy}
\title{Takes an enmtools.species object with presence and background points, and builds
a fitted \code{tidymodels} workflow.}
\usage{
enmtools.tidy(
  species,
  env,
  f = NULL,
  model = "glm",
  test.prop = 0,
  eval = TRUE,
  nback = 1000,
  env.nback = 10000,
  report = NULL,
  overwrite = FALSE,
  rts.reps = 0,
  weights = "equal",
  bg.source = "default",
  verbose = FALSE,
  clamp = TRUE,
  corner = NA,
  bias = NA,
  needs_formula = is.character(model) && model \%in\% c("gam"),
  model_args = list(),
  ...
)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A SpatRaster of environmental data.}

\item{f}{A formula or tidymodels recipe}

\item{model}{A character string specifying the desired model, or a \code{parsnip}
model definition for any model with \code{mode = "classification"}. Default is
"glm". If a character string, choices are the standard ENMTools models:
\code{c("glm", "bc", "dm", "gam", "rf", "rf.ranger", "maxent")}}

\item{test.prop}{Proportion of data to withhold randomly for model evaluation, or "block" for spatially structured evaluation.}

\item{eval}{Determines whether model evaluation should be done.  Turned on by default, but moses turns it off to speed things up.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{env.nback}{Number of points to draw from environment space for environment space discrimination metrics.}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{rts.reps}{The number of replicates to do for a Raes and ter Steege-style test of significance}

\item{weights}{If this is set to "equal", presences and background data will be assigned weights so that the sum of all presence points weights equals the sum of all background point weights.  Otherwise, weights are not provided to the model.}

\item{bg.source}{Source for drawing background points.  If "points", it just uses the background points that are already in the species object.  If "range", it uses the range raster.  If "env", it draws points at randome from the entire study area outlined by the first environmental layer.}

\item{verbose}{Controls printing of various messages progress reports.  Defaults to FALSE.}

\item{clamp}{When set to TRUE, clamps the environmental layers so that predictions made outside the min/max of the training data for each predictor are set to the value for the min/max for that predictor. Prevents the model from extrapolating beyond the min/max bounds of the predictor space the model was trained in, although there could still be projections outside the multivariate training space if predictors are strongly correlated.}

\item{corner}{An integer from 1 to 4.  Selects which corner to use for "block" test data.  By default the corner is selected randomly.}

\item{bias}{An optional raster estimating relative sampling effort per grid cell.  Will be used for drawing background data.}

\item{needs_formula}{Some \code{parsnip} models require a formula to specify the structure of the data. If this is the case, set this to \code{TRUE} so that ENMTools knows to pass the original formula to the \code{tidymodels} workflow. You can safely ignore this if \code{model} is specified as a character string (e.g. a 'classic' ENMTools model)}

\item{model_args}{Arguments to be passed to the \code{parsnip} model specification. For 'classic' ENMTools models specified with a character string argument to \code{model}, these can be found using \code{get_parsnip_model(model)}. Arguments to the underlying engine used by the \code{parsnip} specification can also be specified here. These can be found under the individual model specifications for parsnip engines.}

\item{...}{Additional arguments for future extensions. Not currently used.}
}
\value{
An enmtools model object containing species name, model formula (if any), model object, suitability raster, marginal response plots, and any evaluation objects that were created.
}
\description{
Takes an enmtools.species object with presence and background points, and builds
a fitted \code{tidymodels} workflow.
}
\examples{
data(euro.worldclim)
data(iberolacerta.clade)
enmtools.tidy(iberolacerta.clade$species$monticola, env = euro.worldclim, f = pres ~ bio1 + bio9)
}
