% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.vip.R
\name{enmtools.vip}
\alias{enmtools.vip}
\title{Takes an enmtools.model object, and performs variable importance analyses on it using methods from the vip package}
\usage{
enmtools.vip(
  model,
  metric = "roc_auc",
  nsim = 10,
  method = "permute",
  shap_method = "permshap",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model}{An enmtools.model object}

\item{metric}{The metric to use for measuring how variables affect model predictions}

\item{nsim}{The number of simulations to be run for method "permute"}

\item{method}{A character string or vector containing any combination of "model", "permute", "shap", or "firm".  "model", "permute", and "firm" use functions from the vip package; "shap" computes SHAP values via kernelshap and returns a shapviz object.}

\item{shap_method}{Either "permshap" (default, faster) or "kernelshap" (slower but more accurate for correlated features).  Ignored for enmtools.glm and enmtools.gam models, which always use additive_shap.}

\item{verbose}{Controls printing of messages}

\item{...}{Further arguments to be passed to vip's "vi" functions or kernelshap/permshap, depending on which method is chosen.}
}
\value{
An enmtools.vip object
}
\description{
Takes an enmtools.model object, and performs variable importance analyses on it using methods from the vip package
}
\examples{
\donttest{
#install.extras(repos='http://cran.us.r-project.org')
monticola.glm <- enmtools.glm(iberolacerta.clade$species$monticola,
                              env = euro.worldclim,
                              test.prop = 0.3)
if(check.extras("enmtools.vip")) {
  enmtools.vip(monticola.glm)
}
}
}
