% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.breadth.R
\name{env.breadth}
\alias{env.breadth}
\title{Calculates breadth of a model in environment space using latin hypercube sampling}
\usage{
env.breadth(model, env, tolerance = 0.001, max.reps = 10)
}
\arguments{
\item{model}{An enmtools.model object or a model object that can be projected using the predict() function}

\item{env}{A raster or raster stack of environmental data.}

\item{tolerance}{How close do successive overlap metrics have to be before we decide we're close enough to the final answer}

\item{max.reps}{Maximum number of attempts that will be made to find suitable starting conditions}
}
\description{
Calculates breadth of a model in environment space using latin hypercube sampling
}
