% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.species.hypervolume.R
\name{env.species.hypervolume}
\alias{env.species.hypervolume}
\title{Calculates an environmental hypervolume of presence points using Blonder et al.'s \code{\link[hypervolume]{hypervolume-package}}}
\usage{
env.species.hypervolume(species, env, method = "gaussian",
  use.background = FALSE, standardise = TRUE, reduce_dim = "auto", ...)
}
\arguments{
\item{species}{An enmtools.species object containing presence points}

\item{env}{A raster or raster stack of environmental data.}

\item{method}{Method to use in calculating hypervolume, See \code{\link[hypervolume]{hypervolume}} for details}

\item{use.background}{Should the hypervolume be calulated on background points, instead of presence points?}

\item{standardise}{Should the environmental variables be standardised before calculating the hypervolume?}

\item{reduce_dim}{Should environmental variables be dimension reduced. Currently only option is "auto", which does dimension reduction using principle components analysis, only if the number of environmental variables exceeds a threshold of \code{log(# of observations)}.}

\item{...}{Additional parameters to pass on to the \code{\link[hypervolume]{hypervolume}} function}
}
\description{
Calculates an environmental hypervolume of presence points using Blonder et al.'s \code{\link[hypervolume]{hypervolume-package}}
}
