% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euro.worldclim.R
\docType{data}
\name{euro.worldclim}
\alias{euro.worldclim}
\title{Worldclim data for Europe}
\format{
A raster brick with 19 worldclim layers.
}
\usage{
data(euro.worldclim)
}
\description{
This is a fairly low-resolution set of Worldclim layers for Europe to be used in demonstrating ENMTools functions.
}
\keyword{datasets}
