% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.depends.R
\name{find.extras.missing}
\alias{find.extras.missing}
\title{Find the extra packages needed for an ENMTools function that are currently missing (not available)}
\usage{
find.extras.missing(funs = NULL)
}
\arguments{
\item{funs}{character vector of ENMTools function names to install extra dependencies for. If `NULL`, extras for all ENMTools functions will be installed.}
}
\value{
A character vector with the names of extra packages required by `funs` that are currently not available
}
\description{
Find the extra packages needed for an ENMTools function that are currently missing (not available)
}
\examples{
find.extras.missing("enmtools.calibrate")
}
