% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.depends.R
\name{find.extras}
\alias{find.extras}
\title{Find the extra packages needed for an ENMTools function}
\usage{
find.extras(funs = NULL)
}
\arguments{
\item{funs}{character vector of ENMTools function names to install extra dependencies for. If \code{NULL}, extras for all ENMTools functions will be installed.}
}
\value{
A character vector with the names of extra packages required by \code{funs}
}
\description{
Find the extra packages needed for an ENMTools function
}
\examples{
find.extras("enmtools.calibrate")
}
