% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.overlap.R
\name{geog.range.overlap}
\alias{geog.range.overlap}
\title{Takes two emtools.species objects with range rasters, calculates overlap between them as in Fitzpatrick and Turelli 2006}
\usage{
geog.range.overlap(x, y)
}
\arguments{
\item{x}{An enmtools.species object containing a range raster}

\item{y}{An enmtools.species object containing a range raster}
}
\description{
Takes two emtools.species objects with range rasters, calculates overlap between them as in Fitzpatrick and Turelli 2006
}
