% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{hypervolume_predict}
\alias{hypervolume_predict}
\title{Prediction helper for hypervolume objects}
\usage{
hypervolume_predict(object, newdata)
}
\arguments{
\item{object}{A Hypervolume object from hypervolume package}

\item{newdata}{Data frame or matrix of new points to predict}
}
\value{
Numeric vector of probability density estimates (normalized to max = 1)
}
\description{
Helper function for predicting from hypervolume objects in the tidymodels framework.
Uses hypervolume_estimate_probability for continuous suitability values.
}
