% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iberolacerta.clade.R
\docType{data}
\name{iberolacerta.clade}
\alias{iberolacerta.clade}
\title{GBIF data for a clade of Iberolacerta lizards from spain}
\format{
An enmtools.clade object with seven species and a phylogeny.
}
\usage{
data(iberolacerta.clade)
}
\description{
This is some sample data downloaded from GBIF for the purposes of demonstrating ENMTools functionality.
}
\keyword{datasets}
