% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.reps.R
\name{identity.reps}
\alias{identity.reps}
\title{Takes a set of points, a buffer radius, a sample size, and a mask and returns
randomly sampled points from within that buffer radius.
Code modified from Elith and Hijmans SDM with R tutorial}
\usage{
identity.reps(species, nreps)
}
\arguments{
\item{species}{A list of enmtools.species objects or an enmtools.clade object}

\item{nreps}{Number of replicates to generate}

\item{mask}{A raster to use as a mask}
}
\value{
A data frame of x and y coordinates for pseudoreplicates sampled according to the
identity test of Warren et al. 2008
}
\description{
Takes a set of points, a buffer radius, a sample size, and a mask and returns
randomly sampled points from within that buffer radius.
Code modified from Elith and Hijmans SDM with R tutorial
}
\examples{

}
\keyword{conservatism,}
\keyword{identity}
\keyword{niche}

