% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.plot.R
\name{interactive.plot}
\alias{interactive.plot}
\title{Generic function for making interactive plots of ENMTools models and species}
\usage{
interactive.plot(x, ...)
}
\arguments{
\item{x}{entools.model or species object to plot}

\item{...}{other arguments passed to interactive.plot.enmtools.model or interactive.plot.enmtools.species}
}
\value{
An interactive leaflet plot depicting the predictions and data from the enmtools.model object.
}
\description{
Function that takes an \code{enmtools.model} or \code{enmtools.species}  object and
calls the class-appropriate \code{interactive.plot.xx} function for it.
These functions use \code{\link[leaflet]{leaflet}} for mapping
and will only function properly if you have an active internet connection.
}
