% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.glm.R
\name{make.enmtools.glm}
\alias{make.enmtools.glm}
\title{Function to generate an empty enmtools.glm object and populate it with
user chosen parameters}
\usage{
make.enmtools.glm(env, f = NULL, params)
}
\arguments{
\item{env}{A raster or raster stack of environmental data.}

\item{f}{Standard R formula}
}
\description{
Function to generate an empty enmtools.glm object and populate it with
user chosen parameters
}
