% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makereport.R
\name{makereport}
\alias{makereport}
\title{makereport, creates an html document containing information about an enmtools object}
\usage{
makereport(x, outfile)
}
\arguments{
\item{x}{An enmtools.model, species, test, or clade object}

\item{outfile}{Path to an output file}
}
\description{
makereport, creates an html document containing information about an enmtools object
}
\examples{
makereport(ahli.glm, "./ahli.html")
}
\keyword{enm}
\keyword{html}
\keyword{niche}
\keyword{output}
\keyword{overlap}
\keyword{plot}
\keyword{report}
\keyword{sdm}

