% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{maxnet_bridge}
\alias{maxnet_bridge}
\title{Wrapper function for \code{maxnet::maxnet()}}
\usage{
maxnet_bridge(x, y, regmult = 1, classes = "default", ...)
}
\arguments{
\item{x}{Matrix or data.frame of environmental variables at points}

\item{y}{Single column matrix or data.frame containing a two-level factor}

\item{regmult}{Regularization multiplier passed to maxnet. Default 1.}

\item{classes}{Feature classes for maxnet formula. Default "default".}

\item{...}{Additional arguments passed to maxnet::maxnet()}
}
\value{
The result of calling \code{\link[maxnet:maxnet]{maxnet::maxnet()}}
}
\description{
Wraps \code{\link[maxnet:maxnet]{maxnet::maxnet()}} for use in a \code{parsnip} model specification.
Mostly for internal use. Exported so it works properly with parallel computation in \code{tidymodels}
}
