% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{maxnet_sdm}
\alias{maxnet_sdm}
\title{\code{parsnip} Model specification for maxnet species distribution models}
\usage{
maxnet_sdm(mode = "classification", engine = "maxnet")
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "classification".}

\item{engine}{A single character string specifying what computational engine
to use for fitting. The only engine available is "maxnet".}
}
\description{
\code{parsnip} Model specification for maxnet species distribution models
}
\details{
maxnet fits Maxent models using the glmnet package, providing a
modern implementation of the Maxent algorithm without requiring Java.
}
\examples{
maxnet_sdm()
}
