% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.species.R
\name{merge.species}
\alias{merge.species}
\title{Takes a list of enmtools.species objects and merges them into a single enmtools.species object}
\usage{
\method{merge}{species}(species.list)
}
\arguments{
\item{species.list}{List of enmtools.species objects that you want to merge together}
}
\description{
Takes a list of enmtools.species objects and merges them into a single enmtools.species object
}
