% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moses.list.R
\name{moses.list}
\alias{moses.list}
\title{Takes a list of enmtools.species objects and uses model selection to ask whether they're better treated jointly or separately}
\usage{
moses.list(species.list, env, f = NULL, eval = FALSE, nback = 1000, ...)
}
\arguments{
\item{species.list}{A list of enmtools.species objects, or an enmtools.clade object.}

\item{env}{A raster or raster stack of environmental data.}

\item{nback}{Number of background points to generate, if any}
}
\description{
Takes a list of enmtools.species objects and uses model selection to ask whether they're better treated jointly or separately
}

