% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.variogram.R
\name{multi.variogram}
\alias{multi.variogram}
\title{Takes a raster stack and builds a variogram for each layer.}
\usage{
multi.variogram(env, maxpoints = NA)
}
\arguments{
\item{env}{A raster or raster stack of environmental data.}

\item{maxpoints}{The maximum number of points to use for making your semivariograms. If set to NA uses all points, but this could be quite slow if you have a large number of rasters or high resolution.}
}
\value{
A list containing the variogram data in a data frame as well as a plot of the scaled gammas for each variable.
}
\description{
Takes a raster stack and builds a variogram for each layer.
}
\examples{
multi.variogram(euro.worldclim, maxpoints = 500)
}
