% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.overlap.R
\name{node.overlap}
\alias{node.overlap}
\title{Takes an overlap matrix and a tree and returns average overlap at nodes using Fitzpatrick & Turelli averaging method.  Typicall called via enmtools.aoc rather than used directly.}
\usage{
node.overlap(overlap, tree)
}
\arguments{
\item{overlap}{An overlap matrix}

\item{tree}{A tree}
}
\value{
A data frame of node ages and overlaps
}
\description{
Takes an overlap matrix and a tree and returns average overlap at nodes using Fitzpatrick & Turelli averaging method.  Typicall called via enmtools.aoc rather than used directly.
}
