% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.anw.sdm.R
\name{plot.enmtools.anw.sdm}
\alias{plot.enmtools.anw.sdm}
\title{Plot method for objects of class enmtools.ppmlasso}
\usage{
plot.enmtools.anw.sdm(this.anw.sdm, trans_col = NULL)
}
\arguments{
\item{trans_col}{Transformation to apply to the colour range (Z axis). A character value corresponding to one of \code{ggplot2}'s builtin scale transformations (see argument \code{trans} in \code{\link[ggplot2]{continuous_scale}} for possible choices). \code{trans_col = 'sqrt'} will often work well for ppmlasso plots. If NULL, no transformation is applied.}

\item{this.ppmlasso}{An enmtools.ppmlasso object}
}
\description{
Plot method for objects of class enmtools.ppmlasso
}
