% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.response.R
\name{plot.response}
\alias{plot.response}
\title{plot.response Plots the marginal response of a model to an environmental variable with all other variables held at their mean in env}
\usage{
plot.response(model, env, layer)
}
\arguments{
\item{model}{An enmtools model object}

\item{env}{A RasterLayer or RasterStack object containing environmental data}

\item{layer}{The name of the layer to plot}
}
\value{
results A plot of the marginal response of the model to the environmental variable
}
\description{
plot.response Plots the marginal response of a model to an environmental variable with all other variables held at their mean in env
}
\examples{

}
\keyword{enm,}
\keyword{plot,}
\keyword{response}
\keyword{sdm,}

