% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive.R
\name{plot_interactive.enmtools.species}
\alias{plot_interactive.enmtools.species}
\title{Plot an enmtools.species object on an interactive map}
\usage{
plot_interactive.enmtools.species(x, map_provider = "Esri.WorldPhysical",
  plot_bg = FALSE, cluster_points = FALSE)
}
\arguments{
\item{x}{entools.species object to plot}

\item{map_provider}{Name of a map provider for the underlying interactive base map.
Default is "Esri.WorldPhysical", and attractive topographic map with no place labels. A
preview of all map provider options can be viewed at \url{http://leaflet-extras.github.io/leaflet-providers/preview/}}

\item{plot_bg}{Should background points be plotted?}

\item{cluster_points}{Should points be clustered? If TRUE, points close together
will be grouped into clusters that can be interactively expanded by clicking
on them.}
}
\description{
Function that take an \code{\link{enmtools.species}} object and plots an
interactive map of the presence points, background points (if applicable), and
species range raster (if applicable). This function uses \code{\link{leaflet}} for mapping
and will only function properly if you have an active internet connection.
}
