% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{predict.ppmlasso}
\alias{predict.ppmlasso}
\title{utility functions for ENMTools}
\usage{
\method{predict}{ppmlasso}(object, ..., newdata, interactions = NA)
}
\description{
utility functions for ENMTools
}
