% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{pres_only_sdm}
\alias{pres_only_sdm}
\title{\code{parsnip} Model specification for presence-only species distribution models}
\usage{
pres_only_sdm(mode = "classification", engine = "bioclim")
}
\arguments{
\item{mode}{A single character string for the type of model. The only possible value for this model is "classification".}

\item{engine}{A single character string specifying what computational engine to use for fitting. Possible engines are listed below. The default for this model is "bioclim".}
}
\description{
\code{parsnip} Model specification for presence-only species distribution models
}
\details{
Engines available:
\itemize{
\item \code{\link[dismo:bioclim]{bioclim}}
\item \code{\link[dismo:domain]{domain}}
\item \code{\link[hypervolume:hypervolume]{hypervolume}}
}
}
\examples{
pres_only_sdm()
}
