% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangebreak.ribbon.R
\name{rangebreak.ribbon}
\alias{rangebreak.ribbon}
\title{rangebreak.ribbon Conduct a ribbon rangebreak test as described in Glor and Warren 2011.}
\usage{
rangebreak.ribbon(species.1, species.2, ribbon, env, type, f = NULL,
  width = 1, nreps = 99, nback = 1000, ...)
}
\arguments{
\item{species.1}{An emtools.species object}

\item{species.2}{An enmtools.species object}

\item{ribbon}{An enmtools.species object representing the region of marginal habtiat in the overlap between the species' ranges}

\item{env}{A RasterLayer or RasterStack object containing environmental data}

\item{type}{The type of model to construct, currently accepts "glm", "mx", "bc", or "dm"}

\item{f}{A function to use for model fitting.  Only required for GLM models at the moment.}

\item{width}{Width of the ribbon, in the same units as the occurrence points (e.g, decimal degrees)}

\item{nreps}{Number of replicates to perform}

\item{nback}{Number of background points for models}

\item{...}{Additional arguments to be passed to model fitting functions.}
}
\value{
results A list containing a replicates, models for the empirical data, and summary statistics and plots.
}
\description{
rangebreak.ribbon Conduct a ribbon rangebreak test as described in Glor and Warren 2011.
}
\examples{
data(iberolacerta.clade)
data(euro.worldclim)
cyreni <- iberolacerta.clade$species$cyreni
aranica <- iberolacerta.clade$species$aranica

# We're just going to fudge together occurrence data from a ribbon here
# from random points between the two species' ranges
p <- data.frame(Longitude = runif(50, -4, -2), Latitude = runif(50, 40, 43))
bg <- background.points.buffer(p, 100000, 100, euro.worldclim[[1]])
ribbon <- enmtools.species(species.name = "ribbon", presence.points = p, background.points = bg)

rangebreak.ribbon(cyreni, aranica, ribbon = ribbon, env = euro.worldclim, type = "mx", nreps = 10)

}
\keyword{barrier,}
\keyword{biogeography,}
\keyword{enmtools,}
\keyword{hypothesis}
\keyword{rangebreak,}
\keyword{testing}
