% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.enmtools.model.R
\name{raster.as.im}
\alias{raster.as.im}
\title{Function to convert raster to pixel image
Credit goes to Jeffrey Evans via stack exchange:
https://gis.stackexchange.com/questions/115159/converting-raster-to-im-object-for-point-process-model-covariate-in-r}
\usage{
raster.as.im(im)
}
\description{
Function to convert raster to pixel image
Credit goes to Jeffrey Evans via stack exchange:
https://gis.stackexchange.com/questions/115159/converting-raster-to-im-object-for-point-process-model-covariate-in-r
}
