% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.breadth.R
\name{raster.breadth}
\alias{raster.breadth}
\title{raster.breadth, applies measures of niche breadth to an ENM}
\usage{
raster.breadth(x, verbose = FALSE)
}
\arguments{
\item{x}{A raster file or path to raster file.}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
results A vector containing the two metrics
}
\description{
This function measures the spatial heterogeneity of the distribution of suitability
scores from an ENM.  It returns Levins' two metrics of niche breadth.
}
\examples{
raster.breadth("~/species1.asc")
}
\keyword{keywords}

