% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.cor.R
\name{raster.cor}
\alias{raster.cor}
\title{raster.cor Measure correlation coefficients between two rasters}
\usage{
raster.cor(x, y, method = "spearman")
}
\arguments{
\item{x}{A raster file}

\item{y}{Another raster file}

\item{method}{Determines which method to be used for calculating correlations (from cor function)}
}
\value{
results A correlation coefficient
}
\description{
This function measures similarity in the geographic distribution of values
from two rasters.  Various methods can be specified, but it defaults to
spearman rank correaltion
}
\examples{
raster.cor(env[[1]], env[[2]])

}
\keyword{correlation,}
\keyword{raster}

