% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.overlap.R
\name{raster.overlap}
\alias{raster.overlap}
\title{raster.overlap, measures overlap between two ENMs}
\usage{
raster.overlap(x, y, verbose = FALSE)
}
\arguments{
\item{x}{A raster file}

\item{verbose}{Controls printing of diagnostic messages}

\item{x}{Another raster file}
}
\value{
results A vector containing the two metrics
}
\description{
This function measures similarity in the geographic distribution of suitability scores
from two ENMs.  It returns two metrics, I and D.  These metrics are described in 
Warren et al. 2008.
}
\examples{
raster.overlap("~/species1.asc", "~/species2.asc")
}
\keyword{keywords}

