% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.resid.R
\name{raster.resid}
\alias{raster.resid}
\title{raster.resid Measure correlation coefficients between two rasters}
\usage{
raster.resid(x, y, method = "spearman")
}
\arguments{
\item{x}{A raster file}

\item{y}{Another raster file}
}
\value{
results A raster of residuals from a linear regression between the two supplied rasters
}
\description{
This function measures similarity in the geographic distribution of values
from two rasters.  Various methods can be specified, but it defaults to
spearman rank correaltion
}
\examples{
raster.resid(ahli.raster, allogus.raster)

}
\keyword{correlation,}
\keyword{raster,}
\keyword{residuals}

