% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.standardize.R
\name{raster.standardize}
\alias{raster.standardize}
\title{raster.standardize, standardizes all values in a raster file}
\usage{
raster.standardize(infile = x, outfile = y, verbose = FALSE)
}
\arguments{
\item{verbose}{Controls printing of diagnostic messages}

\item{x}{A raster file in the .asc format, provided as a path}

\item{y}{A path to output a standardized raster file}
}
\description{
This function is used by a lot of the metrics calculated by ENMTools, in order to
standardize suitability scores so they sum to 1 over a geographic space.
}
\examples{
raster.standardize(x)
}
\keyword{keywords}

