% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.sum.R
\name{raster.sum}
\alias{raster.sum}
\title{raster.sum, sums all values in a raster file}
\usage{
raster.sum(x, verbose = FALSE)
}
\arguments{
\item{x}{A raster file}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
sum1 The sum of all suitability scores in the raster
}
\description{
This function is used by a lot of the metrics calculated by ENMTools, in order to
standardize suitability scores so they sum to 1 over a geographic space.
}
\examples{
raster.sum(x)
}
\keyword{keywords}

