% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rasterToPoints2}
\alias{rasterToPoints2}
\title{get terra to produce same output as raster::rasterToPoints}
\usage{
rasterToPoints2(rast)
}
\arguments{
\item{rast}{A raster to convert to points}
}
\description{
get terra to produce same output as raster::rasterToPoints
}
