% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report Build an html report file for an enmtools obect}
\usage{
report(x, path, title, author, date = date(), ...)
}
\arguments{
\item{x}{An object of class enmtools.species, enmtools.clade, enmtools.model, or enmtools.hypothesis.test}

\item{path}{Path for saving the .html report file}

\item{title}{The title to print in the report file}

\item{author}{The author name to print in the report file}

\item{date}{Report generation date - will be automatically populated if not provided}

\item{...}{Extra arguments to be passed to the knit function}
}
\value{
Returns the path to the .html report file
}
\description{
report Build an html report file for an enmtools obect
}
\examples{
report(iberolacerta.clade$species$monticola)

}
\keyword{ENMTools}
\keyword{html}
\keyword{report}
