% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.background.df.R
\name{reps.background.df}
\alias{reps.background.df}
\title{reps.background.csv, creates Monte Carlo replicates for ENM background test using .csv
input files}
\usage{
reps.background.df(x, outfile = FALSE, npoints, reps = 10,
  species = "species", verbose = FALSE, ...)
}
\arguments{
\item{x}{Path to a .csv file to be used for drawing random points}

\item{outfile}{An output .csv file to write replicates to.  If not provided, ENMTools
will return a data frame instead.}

\item{reps}{The number of replicates to create}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
output A data frame containing Monte Carlo replicates to be used in constructing
        models for the background test.
}
\description{
This function generates replicates for the background test, as described in
Warren et al. 2008
}
\examples{
reps.background.df(ahli[,2:3], outfile = "~/myreps.csv", npoints = 10, species = "allogus_east", reps=100)
}
\keyword{keywords}

