% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.background.polygon.R
\name{reps.background.polygon}
\alias{reps.background.polygon}
\title{reps.background.asc, creates Monte Carlo replicates for ENM background test using .asc
input files}
\usage{
reps.background.polygon(x, outfile = FALSE, npoints, reps = 10,
  species = "species", verbose = FALSE, ...)
}
\arguments{
\item{x}{A polygon to use as a mask for drawing points}

\item{outfile}{An output .csv file to write replicates to.  If not provided, ENMTools
will return a data frame instead.}

\item{npoints}{The number of points in each replicate}

\item{reps}{The number of replicates to create}

\item{verbose}{Controls printing of diagnostic messages}

\item{...}{arguments to be passed to the spsample function}
}
\value{
output A data frame containing Monte Carlo replicates to be used in constructing
        models for the background test.
}
\description{
This function generates replicates for the background test, as described in
Warren et al. 2008
}
\examples{
ahli.shape <- circles(ahli[,2:3], d = 5000, lonlat=TRUE)
reps.background.polygon(ahli.shape@polygons, outfile = "myreps.csv", npoints = 20, reps=100)
}
\keyword{keywords}

