% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.background.R
\name{reps.background}
\alias{reps.background}
\title{reps.background, creates Monte Carlo replicates for ENM background test}
\usage{
reps.background(x, outfile = FALSE, npoints, reps = 10,
  species = "species", verbose = FALSE, ...)
}
\arguments{
\item{x}{A raster file, shape file, or data frame to use as a mask for drawing points}

\item{outfile}{An output .csv file to write replicates to.  If not provided, ENMTools
will return a data frame instead.}

\item{npoints}{The number of points in each replicate}

\item{reps}{The number of replicates to create}

\item{verbose}{Controls printing of diagnostic messages}

\item{...}{arguments to be passed to the randomPoints function of dismo}
}
\value{
output A data frame containing Monte Carlo replicates to be used in constructing
        models for the background test.
}
\description{
This function generates replicates for the background test, as described in
Warren et al. 2008
}
\examples{
# Points from a data frame
reps.background(ahli, outfile = "myreps.csv", npoints = 10, reps=100)

# Points from a raster
reps.background(ahli.raster, outfile = "myreps.csv", npoints = 10, reps=100)

#Points from a polygon
ahli.shape <- circles(ahli[,2:3], d = 5000, lonlat=TRUE)
reps.background(ahli.shape, outfile = "myreps.csv", npoints = 10, reps=100)
}
\keyword{keywords}

