% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.identity.R
\name{reps.identity}
\alias{reps.identity}
\title{reps.identity, creates Monte Carlo replicates for ENM identity test}
\usage{
reps.identity(x, outfile = FALSE, reps = 10, verbose = FALSE,
  name.column = 1)
}
\arguments{
\item{x}{One or more .csv files containing species occurrence data}

\item{outfile}{An output .csv file to write replicates to.  If not provided, ENMTools
will return a data frame instead.}

\item{reps}{The number of replicates to create}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
output A data frame containing Monte Carlo replicates to be used in constructing
        models for the identity test.
}
\description{
This function generates replicates for the identity test, as described in
Warren et al. 2008
}
\examples{
reps.identity(x = rbind(ahli, allogus_east), outfile = "myreps.csv", reps=100)
}
\keyword{keywords}

