% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.rangebreak.blob.R
\name{reps.rangebreak.blob}
\alias{reps.rangebreak.blob}
\title{reps.rangebreak.blob, generates Monte Carlo replicates for the blob rangebreak
test of Glor and Warren 2011.}
\usage{
reps.rangebreak.blob(infiles = x, outfile = FALSE, reps = 10,
  verbose = FALSE)
}
\arguments{
\item{infiles}{One or more .csv files of species occurrence data}

\item{outfile}{Path to an output file.  If not provided, replicates will be returned
in a data frame.}

\item{reps}{Number of replicates}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
output A data frame containing replicates for the Monte Carlo test
}
\description{
This function applies blob breaks to the pooled distribution of two species.  Used
in order to construct a null distribution of expected overlaps between allopatric species
for which a linear boundary doesn't make sense.
}
\examples{
reps.rangebreak.blob("~/species_A_and_B.csv", reps=100)
}
\keyword{keywords}

