% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.rangebreak.ribbon.R
\name{reps.rangebreak.ribbon}
\alias{reps.rangebreak.ribbon}
\title{reps.rangebreak.ribbon, generates Monte Carlo replicates for the ribbon rangebreak
test of Glor and Warren 2011.}
\usage{
reps.rangebreak.ribbon(infiles = x, outfile = FALSE, separate = TRUE,
  reps = 10, width = 1, plotty = TRUE, background.raster = NA,
  verbose = FALSE)
}
\arguments{
\item{infiles}{One or more .csv files of species occurrence data}

\item{outfile}{Path to an output file.  If not provided, replicates will be returned
in a data frame.}

\item{separate}{I honestly don't remember what this argument does}

\item{reps}{Number of replicates}

\item{width}{Width of ribbon, in the same units as your input data (i.e., decimal degrees)}

\item{plotty}{Controls plotting of replicates as they are constructed.}

\item{background.raster}{Raster to use as backrgound for plotting replicates}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
output A data frame containing replicates for the Monte Carlo test
}
\description{
This function applies paired linear breaks to the pooled distribution of two species.  Used
in order to construct a null distribution of expected overlaps between species separated
by a ribbon boundary, or between occurrences inside and outside the ribbon.
}
\examples{
reps.rangebreak.ribbon("~/species_A_and_B.csv", reps=100)
}
\keyword{keywords}

