% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.enmtools.model.R
\name{simulate.enmtools.model}
\alias{simulate.enmtools.model}
\title{Simulate a point process from an enmtools.model object}
\usage{
\method{simulate}{enmtools.model}(x, n.points = 1000)
}
\arguments{
\item{x}{entools.model object to simulate from}

\item{n.points}{approximate number of points to simulate. You may get small number fewer or greater.
If you need an exact number of points, generate too many, then drop the extra ones.}
}
\description{
Function that take an \code{enmtools.model} object and simulates points
from it using a point process.
}
