% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.from.file.R
\name{species.from.file}
\alias{species.from.file}
\title{Takes a csv file and uses it to construct one or more enmtools.species objects. It constructs one object per unique value in the "species.col" column.}
\usage{
species.from.file(filename, species.col = "species")
}
\arguments{
\item{filename}{Name of a .csv file}

\item{species.col}{Name of the column from the csv file to use for species names.  Default is "species".}
}
\value{
A list containing species objects, one for each unique value in species.col.
}
\description{
Takes a csv file and uses it to construct one or more enmtools.species objects. It constructs one object per unique value in the "species.col" column.
}
