% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{summary.enmtools.tidy}
\alias{summary.enmtools.tidy}
\title{Summary for objects of class enmtools.glm}
\usage{
\method{summary}{enmtools.tidy}(object, plot = TRUE, ...)
}
\description{
Summary for objects of class enmtools.glm
}
