% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabpfn.utils.R
\name{tabpfn_cache_dir}
\alias{tabpfn_cache_dir}
\title{Get the TabPFN model cache directory}
\usage{
tabpfn_cache_dir()
}
\value{
Character string with the path to the cache directory.
}
\description{
Returns the path to the ENMTools cache directory used for storing
downloaded TabPFN model checkpoints. Creates the directory if it
doesn't exist.
}
