% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabpfn.utils.R
\name{tabpfn_model_path}
\alias{tabpfn_model_path}
\title{Get the local path for a named TabPFN model}
\usage{
tabpfn_model_path(model_name)
}
\arguments{
\item{model_name}{Character string. Name of a finetuned model
(e.g., "sdm-finetuned-nonspatial", "sdm-finetuned-spatial").}
}
\value{
Character string with the path to the model checkpoint file.
}
\description{
Returns the local file path for a named finetuned model. If the model
has not been downloaded yet, it will be downloaded automatically.
}
