% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{tabpfn_predict}
\alias{tabpfn_predict}
\title{Prediction helper for TabPFN objects}
\usage{
tabpfn_predict(object, newdata)
}
\arguments{
\item{object}{A tabpfn_fit object}

\item{newdata}{Data frame or matrix of new points to predict}
}
\value{
Numeric vector of predicted presence probabilities.
}
\description{
Helper function for predicting from TabPFN fit objects in the tidymodels framework.
}
