% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{tabpfn_sdm}
\alias{tabpfn_sdm}
\title{\code{parsnip} Model specification for TabPFN species distribution models}
\usage{
tabpfn_sdm(mode = "classification", engine = "tabpfn")
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "classification".}

\item{engine}{A single character string specifying what computational engine
to use for fitting. The only engine available is "tabpfn".}
}
\description{
\code{parsnip} Model specification for TabPFN species distribution models
}
\details{
TabPFN is a pre-trained tabular foundation model that can be used
for species distribution modeling. Supports local pretrained, local
finetuned, and cloud API backends.
}
