% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.max.R
\name{thin.max}
\alias{thin.max}
\title{reps.rangebreak.ribbon, rarefaction of point data in any number of dimensions.}
\usage{
thin.max(x, cols, npoints)
}
\arguments{
\item{x}{A data frame containing the columns to be used to calculate distances along with whatever other data you need}

\item{cols}{A vector of column names or indices to use for calculating distances}

\item{npoints}{The number of rarefied points to spit out}
}
\value{
output A data frame containing the rarefied point data
}
\description{
reps.rangebreak.ribbon, rarefaction of point data in any number of dimensions.
}
\examples{
thin.max(my.data, c("latitude", "longitude"), 200)
}
\keyword{keywords}

