% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.enm.R
\name{visualize.enm}
\alias{visualize.enm}
\title{visualize.enm, Makes a heatmap of suitability of habitat in environment space according to a given model}
\usage{
visualize.enm(model, env, nbins = 100, layers, plot.points = TRUE)
}
\arguments{
\item{model}{An enmtools.model object}

\item{env}{A set of environmental layers}

\item{nbins}{The number of bins per layer to use for drawing environment space}

\item{layers}{A vector of layer names to use for drawing environment space}

\item{plot.points}{Logical determining whether presence points should be plotted on suitability plot}
}
\value{
suit.plot A two dimensional plot of an ENM
}
\description{
visualize.enm, Makes a heatmap of suitability of habitat in environment space according to a given model
}
\examples{
visualize.enm(1, .001, .001)
}
\keyword{enm}
\keyword{niche}
\keyword{plot}
\keyword{sdm}

