% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.env.R
\name{add.env}
\alias{add.env}
\title{Takes an enmtools.species object and a set of environment layers, and adds the environmental predictor values to the occurrence data.  Typically not called by the user directly.
Code modified from Elith and Hijmans SDM with R tutorial}
\usage{
add.env(species, env, verbose = TRUE)
}
\arguments{
\item{species}{An enmtools.species or enmtools.clade object}

\item{env}{A raster stack of environmental variables}

\item{verbose}{Controls printing of progress messages}
}
\value{
An enmtools.species object with environmental data attached to it.
}
\description{
Takes an enmtools.species object and a set of environment layers, and adds the environmental predictor values to the occurrence data.  Typically not called by the user directly.
Code modified from Elith and Hijmans SDM with R tutorial
}
