% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.bg.R
\name{check.bg.ppmlasso}
\alias{check.bg.ppmlasso}
\title{Takes an emtools.species object and adds background points if they're missing. Looks for range raster first, then goes for environmental layers. This version is for ppmlasso models.}
\usage{
check.bg.ppmlasso(species, env = NA, nback = 1000)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{nback}{Approximate number of background points to generate, if any. Actual number of points generated can be fairly far off due to optimization for speed.}
}
\description{
Takes an emtools.species object and adds background points if they're missing. Looks for range raster first, then goes for environmental layers. This version is for ppmlasso models.
}
