% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.species.R
\name{check.species}
\alias{check.species}
\title{Checking compliance for an object of class enmtools.species.}
\usage{
check.species(this.species)
}
\arguments{
\item{this.species}{An enmtools.species object to be checked.

Checks for existence and proper class of:}

\item{range}{A raster or SpatialPolygon with the actual range they occur in}

\item{presence.points}{A data frame with sampled localities.  If a matrix is passed, it will be cast to a data frame.}

\item{background.points}{A data frame with absence/pseudoabsence/background localities.  If a matrix is passed, it will be cast to a data frame.}

\item{species.name}{A character vector with the species name}

\item{models}{A list of models that are made for the species, which will be stuffed in there as we go along
to pass the check.  This is used by internal enmtools functions to make sure the necessary data is present
before processing anything.}
}
\description{
Checking compliance for an object of class enmtools.species.
}
