% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.species.R
\name{check.species}
\alias{check.species}
\title{Checking compliance for an object of class enmtools.species.}
\usage{
check.species(this.species)
}
\arguments{
\item{range}{A raster or SpatialPolygon with the actual range they occur in}

\item{presence.points}{A data frame with sampled localities}

\item{background.points}{A data frame with absence/pseudoabsence/background localities}

\item{species.name}{A character vector with the species name}

\item{models}{A list of models that are made for the species, which will be stuffed in there as we go along
to pass the check.  This is used by internal enmtools functions to make sure the necessary data is present
before processing anything.}
}
\description{
Checks for existence and proper class of:
}
