% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.species.R
\name{check.species}
\alias{check.species}
\title{Checking compliance for an object of class enmtools.species.}
\usage{
check.species(this.species)
}
\arguments{
\item{this.species}{An enmtools.species object to be checked.}
}
\value{
An enmtools.species object with appropriate formatting.
}
\description{
Checking compliance for an object of class enmtools.species.
}
\examples{
data(iberolacerta.clade)
check.species(iberolacerta.clade$species$monticola)
}
