% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.bc.R
\name{enmtools.bc}
\alias{enmtools.bc}
\title{Takes an emtools.species object with presence and background points, and builds a Bioclim model}
\usage{
enmtools.bc(species, env = NA, test.prop = 0, report = NULL,
  overwrite = FALSE, nback = 1000, rts.reps = 0, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{test.prop}{Proportion of data to withhold randomly for model evaluation, or "block" for spatially structured evaluation.}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{nback}{Number of background points for models. In the case of bioclim models these are only used for evaluation.}

\item{rts.reps}{The number of replicates to do for a Raes and ter Steege-style test of significance}

\item{...}{Arguments to be passed to bioclim()}
}
\description{
Takes an emtools.species object with presence and background points, and builds a Bioclim model
}
\examples{
data(euro.worldclim)
data(iberolacerta.clade)
enmtools.bc(iberolacerta.clade$species$monticola, env = euro.worldclim)
}
