% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.clade.R
\name{enmtools.clade}
\alias{enmtools.clade}
\title{Defining a class for enmtools.clade.  Each clade gets:}
\usage{
enmtools.clade(species = NA, tree = NA)
}
\arguments{
\item{species}{A list of enmtools.species objects}

\item{tree}{A tree showing the relationships between the species}
}
\value{
An enmtools.clade object, either empty or containing a formatted version of the data that was passed into the function.
}
\description{
Defining a class for enmtools.clade.  Each clade gets:
}
