% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.dm.R
\name{enmtools.dm}
\alias{enmtools.dm}
\title{Takes an emtools.species object with presence and background points, and builds a Domain model}
\usage{
enmtools.dm(species, env = NA, test.prop = 0, report = NULL,
  overwrite = FALSE, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{test.prop}{Proportion of data to withhold for model evaluation}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{...}{Arguments to be passed to bioclim()}
}
\description{
Takes an emtools.species object with presence and background points, and builds a Domain model
}

