% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.gam.R
\name{enmtools.gam}
\alias{enmtools.gam}
\title{Takes an emtools.species object with presence and background points, and builds a gam}
\usage{
enmtools.gam(
  species,
  env,
  f = NULL,
  test.prop = 0,
  k = 4,
  nback = 1000,
  env.nback = 10000,
  report = NULL,
  overwrite = FALSE,
  rts.reps = 0,
  weights = "equal",
  gam.method = "REML",
  gam.select = TRUE,
  bg.source = "default",
  verbose = FALSE,
  clamp = TRUE,
  ...
)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{f}{Standard gam formula}

\item{test.prop}{Proportion of data to withhold randomly for model evaluation, or "block" for spatially structured evaluation.}

\item{k}{Dimension of the basis used to represent the smooth term.  See documentation for s() for details.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{env.nback}{Number of points to draw from environment space for environment space discrimination metrics.}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{rts.reps}{The number of replicates to do for a Raes and ter Steege-style test of significance}

\item{weights}{If this is set to "equal", presences and background data will be assigned weights so that the sum of all presence points weights equals the sum of all background point weights.  Otherwise, weights are not provided to the model.}

\item{gam.method}{Defaults to restricted maximum likelihood to facilitate predictor selection, but if you want to use another method you can pass anything here that gam's "method" argument understands.}

\item{gam.select}{Controls whether gam algorithm attempts to optimize smoothness and reduce model complexity.  See help("gam.selection") for details.}

\item{bg.source}{Source for drawing background points.  If "points", it just uses the background points that are already in the species object.  If "range", it uses the range raster.  If "env", it draws points at randome from the entire study area outlined by the first environmental layer.}

\item{verbose}{Controls printing of various messages progress reports.  Defaults to FALSE.}

\item{clamp}{When set to TRUE, clamps the environmental layers so that predictions made outside the min/max of the training data for each predictor are set to the value for the min/max for that predictor. Prevents the model from extrapolating beyond the min/max bounds of the predictor space the model was trained in, although there could still be projections outside the multivariate training space if predictors are strongly correlated.}

\item{...}{Arguments to be passed to gam()}
}
\value{
An enmtools model object containing species name, model formula (if any), model object, suitability raster, marginal response plots, and any evaluation objects that were created.
}
\description{
Takes an emtools.species object with presence and background points, and builds a gam
}
\examples{
\donttest{
data(euro.worldclim)
data(iberolacerta.clade)
if(requireNamespace("mgcv", quietly = TRUE)) {
    enmtools.gam(iberolacerta.clade$species$monticola, env = euro.worldclim, f = pres ~ bio1 + bio9)
}
}
}
