% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.hypervolume.R
\name{enmtools.hypervolume}
\alias{enmtools.hypervolume}
\title{Takes an emtools.species object and environmental layers, and constructs a hypervolume using the R package hypervolume}
\usage{
enmtools.hypervolume(
  species,
  env,
  samples.per.point = 10,
  reduction.factor = 0.1,
  method = "gaussian",
  ...
)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A stack of environmental rasters}

\item{samples.per.point}{To be passed to hypervolume_gaussian}

\item{reduction.factor}{To be passed to hypervolume_project}

\item{method}{Method for constructing hypervolumes, defaults to "gaussian"}

\item{...}{Extra parameters to be passed to hypervolume_gaussian}
}
\description{
Takes an emtools.species object and environmental layers, and constructs a hypervolume using the R package hypervolume
}
\examples{
\dontrun{
# install.packages("hypervolume")
# data(euro.worldclim)
# data(iberolacerta.clade)
# env <- euro.worldclim[[c(1,8,12,17)]]
# monticola.hv <- enmtools.hypervolume(iberolacerta.clade$species$monticola, env = env)
}
}
