% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.maxent.R
\name{enmtools.maxent}
\alias{enmtools.maxent}
\title{Takes an emtools.species object with presence and background points, and builds a maxent model}
\usage{
enmtools.maxent(
  species,
  env,
  test.prop = 0,
  nback = 1000,
  env.nback = 10000,
  report = NULL,
  overwrite = FALSE,
  rts.reps = 0,
  bg.source = "default",
  ...
)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{test.prop}{Proportion of data to withhold randomly for model evaluation, or "block" for spatially structured evaluation.}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{env.nback}{Number of points to draw from environment space for environment space discrimination metrics.}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{rts.reps}{The number of replicates to do for a Raes and ter Steege-style test of significance}

\item{bg.source}{Source for drawing background points.  If "points", it just uses the background points that are already in the species object.  If "range", it uses the range raster.  If "env", it draws points at randome from the entire study area outlined by the first environmental layer.}

\item{...}{Arguments to be passed to maxent()}
}
\value{
An enmtools model object containing species name, model formula (if any), model object, suitability raster, marginal response plots, and any evaluation objects that were created.
}
\description{
Takes an emtools.species object with presence and background points, and builds a maxent model
}
\examples{
\donttest{
install.extras(repos='http://cran.us.r-project.org')
data(euro.worldclim)
data(iberolacerta.clade)
enmtools.maxent(iberolacerta.clade$species$monticola, env = euro.worldclim)
}
}
