% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.maxent.R
\name{enmtools.maxent}
\alias{enmtools.maxent}
\title{Takes an emtools.species object with presence and background points, and builds a maxent model}
\usage{
enmtools.maxent(species, env, test.prop = 0, nback = 1000, report = NULL,
  overwrite = FALSE, rts.reps = 0, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{test.prop}{Proportion of data to withhold for model evaluation}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{rts.reps}{The number of replicates to do for a Raes and ter Steege-style test of significance}

\item{...}{Arguments to be passed to maxent()}
}
\description{
Takes an emtools.species object with presence and background points, and builds a maxent model
}
\examples{
data(euro.worldclim)
data(iberolacerta.clade)
enmtools.maxent(iberolacerta.clade$species$monticola, env = euro.worldclim)
}
