% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.maxent.R
\name{enmtools.maxent}
\alias{enmtools.maxent}
\title{Takes an emtools.species object with presence and background points, and builds a maxent model}
\usage{
\method{print}{enmtools.maxent}(species, env, test.prop = 0, nback = 1000,
  report = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{species}{An enmtools.species object}

\item{env}{A raster or raster stack of environmental data.}

\item{test.prop}{Proportion of data to withhold for model evaluation}

\item{nback}{Number of background points to draw from range or env, if background points aren't provided}

\item{report}{Optional name of an html file for generating reports}

\item{overwrite}{TRUE/FALSE whether to overwrite a report file if it already exists}

\item{...}{Arguments to be passed to maxent()}

\item{formula}{Standard R formula}
}
\description{
Takes an emtools.species object with presence and background points, and builds a maxent model
}
