% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.species.R
\name{enmtools.species}
\alias{enmtools.species}
\title{Defining a class for enmtools.species.}
\usage{
enmtools.species(
  range = NA,
  presence.points = NA,
  background.points = NA,
  species.name = NA,
  models = NA
)
}
\arguments{
\item{range}{A raster or SpatialPolygon with the actual range they occur in}

\item{presence.points}{A data frame with sampled localities}

\item{background.points}{A data frame with absence/pseudoabsence/background localities}

\item{species.name}{A character vector with the species name}

\item{models}{A list of models that are made for the species, which will be stuffed in there as we go along
to pass the check.  This is used by internal enmtools functions to make sure the necessary data is present
before processing anything.}
}
\value{
Returns an enmtools.species object, either empty or populated with the parameter values that were passed into the function.
}
\description{
Each species gets:
}
