% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.overlap.R
\name{env.overlap}
\alias{env.overlap}
\title{Calculates overlap between models in environment space using latin hypercube sampling}
\usage{
env.overlap(model.1, model.2, env, tolerance = 0.001, max.reps = 10,
  cor.method = "spearman")
}
\arguments{
\item{model.1}{An enmtools.model object model object that can be projected using the predict() function}

\item{env}{A raster or raster stack of environmental data.}

\item{tolerance}{How close do successive overlap metrics have to be before we decide we're close enough to the final answer}

\item{max.reps}{Maximum number of attempts that will be made to find suitable starting conditions}

\item{cor.method}{Which method to use for calculating correlations between models}

\item{model.1}{Another enmtools.model object or other model object that can be projected using the predict() function}
}
\description{
Calculates overlap between models in environment space using latin hypercube sampling
}
