% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.tidy.R
\name{hypervolume_bridge}
\alias{hypervolume_bridge}
\title{Wrapper function for \code{hypervolume::hypervolume_gaussian()}}
\usage{
hypervolume_bridge(x, y, method = "gaussian", samples.per.point = 1000, ...)
}
\arguments{
\item{x}{Matrix or data.frame of environmental variables at points}

\item{y}{Single column matrix or data.frame containing a two-level factor}

\item{method}{Method for constructing hypervolumes: "gaussian" (default) or "svm".}

\item{samples.per.point}{Number of random samples per point for hypervolume estimation.}

\item{...}{Additional arguments passed to hypervolume construction function.}
}
\value{
A Hypervolume object with scaling attributes attached
}
\description{
Wraps \code{\link[hypervolume:hypervolume_gaussian]{hypervolume::hypervolume_gaussian()}} for use in a \code{parsnip} model specification.
Mostly for internal use. Exported so it works properly with parallel computation in \code{tidymodels}
}
