% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.test.R
\name{identity.test}
\alias{identity.test}
\title{identity.test Conduct a niche identity/equivalency test as described in Warren et al. 2008.}
\usage{
identity.test(species.1, species.2, env, type, f = NULL, nreps = 99,
  nback = 1000, ...)
}
\arguments{
\item{species.1}{An emtools.species object}

\item{species.2}{An enmtools.species object}

\item{env}{A RasterLayer or RasterStack object containing environmental data}

\item{type}{The type of model to construct, currently accepts "glm", "mx", "bc", "gam", or "dm"}

\item{f}{A function to use for model fitting.  Only required for GLM models at the moment.}

\item{nreps}{Number of replicates to perform}

\item{nback}{Number of background points for models}

\item{...}{Additional arguments to be passed to model fitting functions.}
}
\value{
results A list containing a replicates, models for the empirical data, and summary statistics and plots.
}
\description{
identity.test Conduct a niche identity/equivalency test as described in Warren et al. 2008.
}
\examples{
data(iberolacerta.clade)
data(euro.worldclim)
cyreni <- iberolacerta.clade$species$cyreni
monticola <- iberolacerta.clade$species$monticola
cyreni$range <- background.raster.buffer(cyreni$presence.points, 100000, euro.worldclim)
monticola$range <- background.raster.buffer(monticola$presence.points, 100000, euro.worldclim)
identity.test(cyreni, monticola, env = euro.worldclim, type = "mx", nreps = 10)
}
\keyword{enmtools,}
\keyword{equivalency,}
\keyword{hypothesis}
\keyword{identity,}
\keyword{testing}
