% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.test.R
\name{identity.test}
\alias{identity.test}
\title{identity.test Measure standardized residuals from a linear regression between two rasters.}
\usage{
identity.test(species.1, species.2, env, type, f = NULL, nreps = 99, ...)
}
\arguments{
\item{species.1}{An emtools.species object}

\item{species.2}{An enmtools.species object}

\item{env}{A RasterLayer or RasterStack object containing environmental data}

\item{type}{The type of model to construct, currently accepts "glm", "mx", "bc", or "dm"}

\item{f}{A function to use for model fitting.  Only required for GLM models at the moment.}

\item{nreps}{Number of replicates to perform}

\item{...}{Additional arguments to be passed to model fitting functions.}
}
\value{
results A list containing a replicates, models for the empirical data, and summary statistics and plots.
}
\description{
This function builds a linear model for the relationship between two rasters, and returns the linear model
and a spatial raster of the residuals from that model.
}
\examples{
identity.test(ahli, allogus, env, type = "glm", f = layer.1 + layer.2 + layer.3, nreps = 10, ...)

}
\keyword{enmtools,}
\keyword{equivalency,}
\keyword{hypothesis}
\keyword{identity,}
\keyword{testing}

