% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.depends.R
\name{install.extras}
\alias{install.extras}
\title{Convenience function that installs all extra packages used in ENMTools}
\usage{
install.extras()
}
\description{
Convenience function that installs all extra packages used in ENMTools
}
