% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.depends.R
\name{install.extras}
\alias{install.extras}
\title{install.extras}
\usage{
install.extras(repos = "http://cran.us.r-project.org", ...)
}
\arguments{
\item{repos}{URL for the repository to use for installing R packages}

\item{...}{Other parameters to be passed to \code{install.packages}}
}
\value{
No return value.
}
\description{
Convenience function that installs all extra packages used in ENMTools.  ENMTools uses
functions from a lot of external packages, and due to CRAN best practices it doesn't install
those packages by default.  The function install.extras() just calls install.packages with a
list of all of the extra packages.  At present this list includes mgcv, ecospat, randomForest,
hypervolume, ape, ppmlasso, leaflet, ranger, CalibratR, caret, and ResourceSelection.
}
\examples{
\donttest{
install.extras()
}
}
