% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmtools.ppmlasso.R
\name{make.enmtools.ppmlasso}
\alias{make.enmtools.ppmlasso}
\title{Function to generate an empty enmtools.ppmlasso object and populate it with
user chosen parameters}
\usage{
make.enmtools.ppmlasso(env, f = NULL, params)
}
\arguments{
\item{env}{A raster or raster stack of environmental data.}

\item{f}{Standard R formula}

\item{params}{A matrix of parameters for the model. Must be a vector with length equal to the number
of terms in the formula f. If f is NULL a formula is built automatically with an intercept, two polynomial terms
per environmental variable plus their pairwise interactions, which gives 1 + n.env\*2 + (n.env \* (n.env - 1) / 2)}
}
\description{
Function to generate an empty enmtools.ppmlasso object and populate it with
user chosen parameters
}
