% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.cor.R
\name{raster.cor}
\alias{raster.cor}
\title{Calculates the correlation coefficient between two rasters.}
\usage{
raster.cor(x, y, method = "spearman")
}
\arguments{
\item{x}{A raster representing the spatial projection of an SDM, or an ENMTools model object with a suitability raster.}

\item{y}{A raster representing the spatial projection of an SDM, or an ENMTools model object with a suitability raster.}

\item{method}{The method to be used for calculating correlations.  Defaults to spearman, but can take "kendall" or "pearson" as well.}
}
\description{
Calculates the correlation coefficient between two rasters.
}

