% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.pca.R
\name{raster.pca}
\alias{raster.pca}
\title{raster.pca, PCA on a set of environmental rasters}
\usage{
raster.pca(env, n)
}
\arguments{
\item{env}{A set of environmental layers}

\item{n}{The number of PCA layers to return}
}
\value{
A list containing a stack of rasters representing the top n pca axes of the initial environmental variables, as well as the pca object from the analysis that produced them.
}
\description{
raster.pca, PCA on a set of environmental rasters
}
\examples{
env.pca <- raster.pca(euro.worldclim, 2)
}
\keyword{environment}
\keyword{pca}
\keyword{raster}
