% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.standardize.R
\name{raster.standardize}
\alias{raster.standardize}
\title{raster.standardize, standardizes all values in a raster file}
\usage{
raster.standardize(x, verbose = FALSE)
}
\arguments{
\item{x}{A raster or RasterLayer object, or an ENMTools model object containing a suitability raster.}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
A new raster object, standardized so that values sum to 1
}
\description{
This function is used by a lot of the metrics calculated by ENMTools, in order to
standardize suitability scores so they sum to 1 over a geographic space.
}
\examples{
data(euro.worldclim)
raster.standardize(euro.worldclim[[1]])
}
\keyword{keywords}
