% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.enm.R
\name{visualize.enm}
\alias{visualize.enm}
\title{visualize.enm, Makes a heatmap of suitability of habitat in environment space according to a given model}
\usage{
visualize.enm(model, env, nbins = 100, layers = names(env)[1:2],
  plot.test.data = FALSE, plot.points = TRUE)
}
\arguments{
\item{model}{An enmtools.model object}

\item{env}{A set of environmental layers}

\item{nbins}{The number of bins per layer to use for drawing environment space}

\item{layers}{A vector of layer names to use for drawing environment space}

\item{plot.test.data}{Logical determining whether test data should be plotted, if present.  If test data is plotted, it will appear as translucent green triangles.}

\item{plot.points}{Logical determining whether presence points should be plotted on suitability plot}
}
\value{
suit.plot A two dimensional plot of an ENM
}
\description{
visualize.enm, Makes a heatmap of suitability of habitat in environment space according to a given model
}
\examples{
data(iberolacerta)
data(euro.worldclim)
aurelioi.mx <- enmtools.maxent(iberolacerta.clade$species$aurelioi, euro.worldclim)
visualize.enm(aurelioi.mx,euro.worldclim, layers = c("bio14", "bio13"))
}
\keyword{enm}
\keyword{niche}
\keyword{plot}
\keyword{sdm}
